+++ src/KSeFCli/IWithConfigCommand.cs
@@ -94,17 +94,13 @@ public abstract class IWithConfigCommand : IGlobalCommand
         return new TokenStore.Key(config.ActiveProfile, profile);
     }
 
-    private static string StatusInfoToString(StatusInfo statusInfo)
+    private static string StatusInfoToString(OperationStatusInfo statusInfo)
     {
         StringBuilder sb = new StringBuilder();
         sb.Append($"Code: {statusInfo.Code}, Description: {statusInfo.Description}");
         if (statusInfo.Details != null && statusInfo.Details.Any())
         {
             sb.Append($", Details: [{string.Join(", ", statusInfo.Details)}]");
-        }
-        if (statusInfo.Extensions != null && statusInfo.Extensions.Any())
-        {
-            sb.Append($", Extensions: {{{string.Join(", ", statusInfo.Extensions.Select(kv => $"{kv.Key}: {kv.Value}"))}}}");
         }
         return sb.ToString();
     }